(bluefox) working on tests

+++ test/lib/setup.js
@@ -283,25 +283,46 @@ function installJsController(cb) {
             // check if port 9000 is free, else admin adapter will be added to running instance
             var client = new require('net').Socket();
             client.connect(9000, '127.0.0.1', function() {
-                console.error('One instance of application is running on this PC');
+                console.error('Cannot initiate fisrt run of test, because one instance of application is running on this PC. Stop it and repeat.');
                 process.exit(0);
             });
 
             setTimeout(function () {
                 client.destroy();
-                console.log('installJsController: no js-controller => install from git');
+                if (!fs.existsSync(rootDir + 'tmp/node_modules/' + appName + '.js-controller')) {
+                    console.log('installJsController: no js-controller => install from git');
 
-                child_process.execSync('npm install https://github.com/' + appName + '/' + appName + '.js-controller/tarball/master --prefix ./  --production', {
-                    cwd:   rootDir + 'tmp/',
-                    stdio: [0, 1, 2]
-                });
+                    child_process.execSync('npm install https://github.com/' + appName + '/' + appName + '.js-controller/tarball/master --prefix ./  --production', {
+                        cwd:   rootDir + 'tmp/',
+                        stdio: [0, 1, 2]
+                    });
+                } else {
+                    console.log('Setup js-controller...');
+                    var __pid;
+                    if (debug) {
+                        // start controller
+                        child_process.exec('node ' + appName + '.js setup first', {
+                            cwd: rootDir + 'tmp/node_modules/' + appName + '.js-controller',
+                            stdio: [0, 1, 2]
+                        });
+                    } else {
+                        child_process.fork(appName + '.js', ['setup', 'first'], {
+                            cwd:   rootDir + 'tmp/node_modules/' + appName + '.js-controller',
+                            stdio: [0, 1, 2]
+                        });
+                    }
+                }
 
                 // let npm install admin and run setup
                 checkIsControllerInstalled(function () {
-                    var _pid = child_process.fork(appName + '.js', ['stop'], {
-                        cwd:   rootDir + 'node_modules/' + appName + '.js-controller',
-                        stdio: [0, 1, 2]
-                    });
+                    var _pid;
+
+                    if (fs.existsSync(rootDir + 'node_modules/' + appName + '.js-controller/' + appName + '.js')) {
+                        _pid = child_process.fork(appName + '.js', ['stop'], {
+                            cwd:   rootDir + 'node_modules/' + appName + '.js-controller',
+                            stdio: [0, 1, 2]
+                        });
+                    }
 
                     waitForEnd(_pid, function () {
                         // change ports for object and state DBs
@@ -378,17 +399,18 @@ function setupController(cb) {
 
 function startAdapter(objects, states, callback) {
     console.log('startAdapter...');
-    if (fs.existsSync(rootDir + 'tmp/node_modules/' + pkg.name + '/' + pkg.main)) {
+
+    if (fs.existsSync(rootDir + 'tmp/node_modules/' + pkg.name + '/' + (pkg.main || 'main.js'))) {
         try {
             if (debug) {
                 // start controller
-                pid = child_process.exec('node node_modules/' + pkg.name + '/' + pkg.main, {
+                pid = child_process.exec('node node_modules/' + pkg.name + '/' + (pkg.main || 'main.js'), {
                     cwd: rootDir + 'tmp',
                     stdio: [0, 1, 2]
                 });
             } else {
                 // start controller
-                pid = child_process.fork('node_modules/' + pkg.name + '/' + pkg.main, {
+                pid = child_process.fork('node_modules/' + pkg.name + '/' + (pkg.main || 'main.js'), {
                     cwd:   rootDir + 'tmp',
                     stdio: [0, 1, 2]
                 });
@@ -397,12 +419,18 @@ function startAdapter(objects, states, callback) {
             console.error(JSON.stringify(error));
         }
     } else {
-        console.error('Cannot find: ' + rootDir + 'tmp/node_modules/' + pkg.name + '/' + pkg.main);
+        console.error('Cannot find: ' + rootDir + 'tmp/node_modules/' + pkg.name + '/' + (pkg.main || 'main.js'));
     }
     if (callback) callback(objects, states);
 }
 
-function startController(callback) {
+function startController(isStartAdapter, callback) {
+    if (typeof isStartAdapter === 'function') {
+        callback = isStartAdapter;
+        isStartAdapter = true;
+    }
+    if (isStartAdapter === undefined) isStartAdapter = true;
+
     if (pid) {
         console.error('Controller is already started!');
     } else {
@@ -439,7 +467,11 @@ function startController(callback) {
                 isObjectConnected = true;
                 if (isStatesConnected) {
                     console.log('startController: started!');
-                    startAdapter(objects, states, callback);
+                    if (isStartAdapter) {
+                        startAdapter(objects, states, callback);
+                    } else {
+                        callback && callback(objects, states);
+                    }
                 }
             }
         });
@@ -479,7 +511,7 @@ function startController(callback) {
     }
 }
 
-function stopController(cb) {
+function stopAdapter(cb) {
     if (!pid) {
         console.error('Controller is not running!');
         if (cb) {
@@ -488,87 +520,73 @@ function stopController(cb) {
             }, 0);
         }
     } else {
-        var timeout;
-        if (objects) {
-            console.log('Set system.adapter.' + pkg.name + '.0');
-            objects.setObject('system.adapter.' + pkg.name + '.0', {
-                common:{
-                    enabled: false
-                }
-            });
-        }
-
         pid.on('exit', function (code, signal) {
             if (pid) {
                 console.log('child process terminated due to receipt of signal ' + signal);
-                if (timeout) {
-                    clearTimeout(timeout);
-                    timeout = null;
-                }
-                if (objects) {
-                    objects.destroy();
-                    objects = null;
-                }
-                if (states) {
-                    states.destroy();
-                    states = null;
-                }
-
-                if (cb) {
-                    cb(true);
-                    cb = null;
-                }
+                cb && cb();
                 pid = null;
             }
         });
 
         pid.on('close', function (code, signal) {
             if (pid) {
-                if (timeout) {
-                    clearTimeout(timeout);
-                    timeout = null;
-                }
-                console.log('child process terminated due to receipt of signal ' + signal);
-
-                if (objects) {
-                    objects.destroy();
-                    objects = null;
-                }
-                if (states) {
-                    states.destroy();
-                    states = null;
-                }
-
-                if (cb) {
-                    cb(true);
-                    cb = null;
-                }
+                cb && cb();
                 pid = null;
             }
         });
 
         pid.kill('SIGTERM');
-
-        timeout = setTimeout(function () {
-            timeout = null;
-            console.log('child process NOT terminated');
-
-            if (objects) {
-                objects.destroy();
-                objects = null;
-            }
-            if (states) {
-                states.destroy();
-                states = null;
-            }
-
-            if (cb) {
-                cb(false);
-                cb = null;
-            }
-            pid = null;
-        }, 5000);
     }
+}
+
+function _stopController() {
+    if (objects) {
+        objects.destroy();
+        objects = null;
+    }
+    if (states) {
+        states.destroy();
+        states = null;
+    }
+}
+
+function stopController(cb) {
+    var timeout;
+    if (objects) {
+        console.log('Set system.adapter.' + pkg.name + '.0');
+        objects.setObject('system.adapter.' + pkg.name + '.0', {
+            common:{
+                enabled: false
+            }
+        });
+    }
+
+    stopAdapter(function () {
+        if (timeout) {
+            clearTimeout(timeout);
+            timeout = null;
+        }
+
+        _stopController();
+
+        if (cb) {
+            cb(true);
+            cb = null;
+        }
+    });
+
+    timeout = setTimeout(function () {
+        timeout = null;
+        console.log('child process NOT terminated');
+
+        _stopController();
+
+        if (cb) {
+            cb(false);
+            cb = null;
+        }
+        pid = null;
+    }, 5000);
 }
 
 // Setup the adapter
@@ -593,4 +611,7 @@ if (typeof module !== undefined && module.parent) {
     module.exports.startController  = startController;
     module.exports.stopController   = stopController;
     module.exports.setupController  = setupController;
+    module.exports.stopAdapter      = stopAdapter;
+    module.exports.startAdapter     = startAdapter;
+    module.exports.appName          = appName;
 }
